requirements/test.txt
pytest-celery~=0.0.0
pytest~=6.2
coverage
pytest-timeout
